Refactor creditCard
